V1.1.22 - extra tests + reduce pressure
[TEST]
- added tests for dao's (partyDao doesn't pass yet)
- updated MainActivity Test

reduce pressure: because this is my first app and since I've gone overboard in what I'm trying to make as a first app in one semester, I've decided that the product which I'm working towards is only an mvp. In my spare time I'll improve upon this app and this idea. This means that some functionality which would be expected is being left out:
- party detail overview
- the crud for games
- update and delete parties

Extra features are planned, just not for the school project